# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pelaporan keuangan dengan manajemen users, transaksi, dan master data",
        "title": "Pelaporan Keuangan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/jenis-pembayaran": {
            "get": {
                "description": "Get all payment types with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-JenisPembayaran"
                ],
                "summary": "Get all payment types",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all payment types success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseGetAllSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.ResJenisPembayaran"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/definitions/helpers.Pagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid pagination data",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "No payment types found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new payment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-JenisPembayaran"
                ],
                "summary": "Create new payment type",
                "parameters": [
                    {
                        "description": "Payment type data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InputJenisPembayaran"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create payment type success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid input data or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/jenis-pembayaran/{id}": {
            "get": {
                "description": "Get payment type details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-JenisPembayaran"
                ],
                "summary": "Get payment type details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get payment type detail success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseGetDetailSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.ResJenisPembayaran"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid payment type ID",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Payment type not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing payment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-JenisPembayaran"
                ],
                "summary": "Update payment type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update payment type data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InputJenisPembayaran"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update payment type success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid payment type ID or request data",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Payment type not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific payment type by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-JenisPembayaran"
                ],
                "summary": "Delete payment type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment Type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete payment type success",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad request - Invalid payment type ID",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Payment type not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dtos.InputJenisPembayaran": {
            "type": "object",
            "required": [
                "nama"
            ],
            "properties": {
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ResJenisPembayaran": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "helpers.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "next_page": {
                    "type": "integer",
                    "example": 2
                },
                "page_size": {
                    "type": "integer",
                    "example": 5
                },
                "previous_page": {
                    "type": "integer",
                    "example": 0
                },
                "total_data": {
                    "type": "integer",
                    "example": 50
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "helpers.ResponseCUDSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "helpers.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "helpers.ResponseGetAllSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "pagination": {
                    "$ref": "#/definitions/helpers.Pagination"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "helpers.ResponseGetDetailSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}